requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
ecdsa==0.19.1
fastapi==0.115.0
google-ai-generativelanguage==0.6.10
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.71.0
gotrue==2.12.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
packaging==25.0
passlib==1.7.4
postgrest==0.18.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.6.1
pydantic_core==2.41.4
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.3
realtime==2.22.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.38.6
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
